refactor: extract plugin path normalizer
